+++ compiler/src/iree/compiler/DispatchCreation/BUILD.bazel
@@ -48,6 +48,7 @@ iree_compiler_cc_library(
         "RemoveTensorBarriers.cpp",
         "SetEncoding.cpp",
         "SetSplitReductionSizes.cpp",
+        "SinkBitCast.cpp",
         "SinkReshapes.cpp",
         "SplitReduction.cpp",
         "TensorPadToTensorInsertSlice.cpp",
